Move pi command to a dedicated file
